Finish the main functionalities
