+++ src/components/App.js
@@ -4,6 +4,7 @@ import PropType from 'prop-types';
 import { connect } from 'react-redux';
 import LogIn from './auth/LogIn';
 import Dashboard from './Dashboard';
+import FavoriteList from './FavoriteList';
 import NavBar from './NavBar';
 import Home from './Home';
 import { isLoggedIn } from '../actions';
@@ -19,6 +20,7 @@ const App = ({ isLoggedIn }) => {
         <Route path="/" component={Home} exact />
         <Route path="/login" component={LogIn} exact />
         <Route path="/dashboard" component={Dashboard} exact />
+        <Route path="/favorites" component={FavoriteList} exact />
       </Switch>
     </Router>
   );
