+++ src/components/constans.js
@@ -1,0 +1,29 @@
+const animeList = [
+  {
+    id: 1,
+    title: 'Fullmetal Alchemist: Brotherhood',
+    imageUrl: 'https://cdn.myanimelist.net/images/anime/1223/96541.jpg',
+  },
+  {
+    id: 2,
+    title: 'Shingeki no Kyojin: The Final Season',
+    imageUrl: 'https://cdn.myanimelist.net/images/anime/1000/110531.jpg',
+  },
+  {
+    id: 3,
+    title: 'Steins;Gate',
+    imageUrl: 'https://cdn.myanimelist.net/images/anime/5/73199.jpg',
+  },
+  {
+    id: 4,
+    title: 'Hunter x Hunter (2011)',
+    imageUrl: 'https://cdn.myanimelist.net/images/anime/11/33657.jpg',
+  },
+  {
+    id: 5,
+    title: 'Mob Psycho 100 II',
+    imageUrl: 'https://cdn.myanimelist.net/images/anime/1918/96303.jpg',
+  },
+];
+
+export default animeList;
